[Bevy0.6対応]main.rsだけコンパイル成功するよう修正
